Change type (float to decimal) for ValueInCent property

+++ GivingBackChange/GivingBackChange.Business/BusinessObjects/CoinBo.cs
@@ -2,7 +2,7 @@
 {
     public class CoinBo
     {
-        public CoinBo(int id, string label, float value, int quantity)
+        public CoinBo(int id, string label, decimal value, int quantity)
         {
             this.Id = id;
             this.Label = label;
@@ -14,7 +14,7 @@
 
         public string Label { get; set; }
 
-        public float Value { get; set; }
+        public decimal Value { get; set; }
 
         public int ValueInCent => (int)(this.Value * 100);
